requirements.txt
apscheduler==3.11.0
blinker==1.9.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
flask==2.2.5
flask-apscheduler==1.13.1
idna==3.10
itsdangerous==2.2.0
jinja2==3.1.4
markupsafe==3.0.2
mysql-connector-python==9.1.0
pip @ file:///private/var/folders/c_/qfmhj66j0tn016nkx_th4hxm0000gp/T/abs_0ato15bisz/croot/pip_1723484599651/work
prometheus-client==0.21.0
prometheus-flask-exporter==0.23.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
redis==5.2.0
requests==2.32.3
setuptools==75.1.0
six==1.16.0
tzlocal==5.2
urllib3==2.2.3
uv==0.5.4
werkzeug==3.1.3
wheel==0.44.0
